Remove getElementForStep in favor of step.target
