+++ src/js/utils/cleanup.js
@@ -1,6 +1,3 @@
-import { preventModalBodyTouch } from './modal.jsx';
-import { getElementForStep } from './dom';
-
 /**
  * Cleanup the steps and set pointerEvents back to 'auto'
  * @param tour The tour object
@@ -11,7 +8,7 @@ export function cleanupSteps(tour) {
 
     steps.forEach((step) => {
       if (step.options && step.options.canClickTarget === false && step.options.attachTo) {
-        const stepElement = getElementForStep(step);
+        const stepElement = step.target;
 
         if (stepElement instanceof HTMLElement) {
           stepElement.style.pointerEvents = 'auto';
