+++ test/unit/utils/dom.spec.js
@@ -1,9 +1,21 @@
 import {
-  elementIsHidden,
-  getElementForStep
+  cleanupStepEventListeners,
+  elementIsHidden
 } from '../../../src/js/utils/dom';
 
 describe('DOM Utils', function() {
+  describe('cleanupStepEventListeners', function() {
+    it('removes listeners', () => {
+      const mock = {
+        _onScreenChange() {}
+      };
+
+      cleanupStepEventListeners.call(mock);
+
+      expect(mock._onScreenChange).toBe(null);
+    });
+  });
+
   describe('elementIsHidden', function() {
     it('returns true when hidden', () => {
       const element = document.createElement('div');
@@ -11,39 +23,6 @@ describe('DOM Utils', function() {
       element.setAttribute('offsetWidth', 0);
 
       expect(elementIsHidden(element), 'evaluates to true when offset height and width are 0').toBeTruthy();
-    });
-  });
-
-  describe('getElementForStep', function() {
-    it('attachTo object - element is HTMLElement', () => {
-      const element = document.createElement('div');
-      const step = {
-        options: {
-          attachTo: {
-            element,
-            on: 'bottom'
-          }
-        }
-      };
-
-      expect(getElementForStep(step), 'returns element as is for HTMLElement').toEqual(element);
-    });
-
-    it('attachTo object - element is selector', () => {
-      const element = document.createElement('div');
-      element.classList.add('foo');
-      document.body.appendChild(element);
-
-      const step = {
-        options: {
-          attachTo: {
-            element: '.foo',
-            on: 'bottom'
-          }
-        }
-      };
-
-      expect(getElementForStep(step), 'returns element from selector').toEqual(element);
     });
   });
 });
